Improve method of avoiding cxi_send_lat de-sync
The client and server side must perform the same number of iterations,
so some special handling is needed when the "duration" option is used.
A previous fix worked around this with a tradeoff: a significant
increase in the overhead of running each iteration. This commit replaces
that fix with a new one that does not increase run time.

Signed-off-by: Andrew Gailey <andrew.gailey@hpe.com>