Fix unitialized unions (and structs)
gcc>=15 and clang>=18 only initialize the first member in a union when
using `{ 0 }` as initializer. This leads to errors like the following
when starting any cxi_*_* tool:

    Received message (1600747619) larger than buf (1020)
    Failed to exchange client/server config: Message too long

or

    Timed out waiting for LINK event
    Failed to get LINK event: Timer expired

Running the tools through valgrind then reveal the use of uninitalized
data.

In libfabric and cxi-driver (test code) this has been resolved by
implicitly relying on C23 and using its `{}` initializers instead.
Hence, following that pattern here as well.

The Linux kernel instead passes `-fzero-init-padding-bits=unions` to
restore old GCC behaviour for now.

A different solution would be something like this:

    #if __STDC__VERSION__ >= 202311L || __GNUC__ >= 15 || __clang_major__ >= 18
    #define LIBCXI_EMPTY_INITIALIZER(...) {}
    #else
    #warning not using empty initializers
    #define LIBCXI_EMPTY_INITIALIZER(...) __VA_ARGS__
    #endif

as suggested here:
https://github.com/Mbed-TLS/TF-PSA-Crypto/pull/482/commits/c73e107a17fc968a9f96326ffb85bf5d5b02de40

Signed-off-by: Tiziano Müller <tiziano.mueller@hpe.com>